rename fixture dir
