+++ tests/Issues/ExtendsNotReadOnlyClass/Source/ParentIsNonReadOnlyClass.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\Core\Tests\Issues\ExtendsNonReadOnlyClass\Source;
+namespace Rector\Core\Tests\Issues\ExtendsNotReadOnlyClass\Source;
 
 class ParentIsNonReadOnlyClass
 {
