+++ tests/Issues/ExtendsNotReadOnlyClass/config/configured_rule.php
